% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dame.Function.R
\name{Dame0.Function}
\alias{Dame0.Function}
\title{Function to simuulate spatial process for shot chart data for Damian Lillard}
\usage{
Dame0.Function(sigma.sq.T, tau.sq.T, phi.T, grid)
}
\arguments{
\item{sigma.sq.T}{"True" value for sigma.sq}

\item{tau.sq.T}{"True" value for tau.sq}

\item{phi.T}{"True" value for phi}

\item{grid}{Length of one side of the grid assocciated with the simulated spatial process}
}
\description{
This function allows you to explore differnt "true" parameter values assocciated with the simulated spatial process.
}
\examples{
Spatial.Function()
}
\keyword{Spatial}
